pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MoYunTengCli"
version = "0.1.1"
description = "CLI and Python package for Moyunteng box VM management"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  {name = "OpenAI Codex"},
]
keywords = ["moyunteng", "myt", "android", "cli", "sdk"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: System :: Systems Administration",
]
dependencies = [
  "requests>=2.28.0,<3",
  "PyYAML>=6.0,<7",
]

[project.urls]
Homepage = "https://github.com/Wowgreat/MoYunTengCli"
Repository = "https://github.com/Wowgreat/MoYunTengCli"
Issues = "https://github.com/Wowgreat/MoYunTengCli/issues"

[project.scripts]
myt-cli = "myt_cli.cli:main"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["myt_cli*"]
